rebase master

+++ src/main/kotlin/com/github/anddd7/repository/ProductRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface ProductRepository : ReactiveCrudRepository<Product, Long>
+interface ProductRepository : ReactiveCrudRepository<Product, Int>
